algo-trading-bot/pyproject.toml
[project]
name = "algo-trading-bot"
version = "0.1.0"
description = "Algo trading bot for crypto markets with WebSocket-first APIs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Trading Team" }
]

dependencies = [
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "sqlite-utils>=3.36",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "torch>=2.2.0",
    "ta-lib>=0.4.28",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
